Basic Lantern (that can flicker)
Added a basic lantern prefab with a point light 2D that observes shadows. This light has a script that allows it to flicker (customizable).

NOTE: In order for the script to be able to access the Light 2D component, a line must be added to the top of the Light2D.cs file in \Library\PackageCache\com.unity.render-pipelines.universal@10.6.0\Runtime\2D\Light2D.cs
This line is "using UnityEngine.Experimental.Rendering.Universal;", according to https://stackoverflow.com/questions/58120976/how-can-i-access-the-unity-light-2d-script-component-via-script