Able to define "Scenes" for which the Camera will follow
Scene objects determine different camera tracking behavior for a given readion. They must added to the list of scenes under the main camera, and are defined by min and max x and y values, as well as the type of Scene (in an enum) and it's activity status. Inactive scenes are ignored.